Updated main activity (failure to end goals)
After converting PCM 16 bit to float,
and then hearing the horrible sound quality in 512 buffer, it was obvious that realtime filtering using the LPRez filtering function from the Beads library will not work.  That part  of the project has ended and I will do complicated modulations with audio with a DAW conjoined with hacking properites.  I will now focus on RESTful API protocols with java.  May use a service for calling the GCD cloud but not sure if it is totally neccesary.